- Modification du ModifEmployeController  -> Ajout de la liste des emplois precedents de l'employe à la dataMap de referenceData
- Modification de add_modif_employe.jsp
 -> Ajout de la liste des emplois precedents
- Ajout du ListeEmploiPrecedentsDecorator

- Creation de la jsp add_modif_emploi_precedent.jsp.
- Creation du AddEmploiPrecedentController.
- Creation du ModifEmploiPrecedentController.
- Creation du SupprEmploiPrecedentController.
- Creation du AddModifEmploiPrecedentValidator.
- Mise a jour de spring-servlet.xml :
 -> Ajout du bean add_emploi_precedent.htm.
 -> Ajout du bean modif_emploi_precedent.htm.
 -> Ajout du bean suppr_emploi_precedent.htm.

+++ src/entities/EmploiPrecedent.java
@@ -3,7 +3,7 @@ package entities;
 import static javax.persistence.GenerationType.SEQUENCE;
 
 import java.io.Serializable;
-import java.sql.Date;
+import java.util.Date;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -21,7 +21,7 @@ public class EmploiPrecedent implements Serializable {
 	@Column(name="ID")
 	@GeneratedValue(strategy=SEQUENCE, generator = "SEQ_EMPLOI_PRECEDENT")
 	@SequenceGenerator(name="SEQ_EMPLOI_PRECEDENT", sequenceName = "SEQ_EMPLOI_PRECEDENT")
-	private long id;
+	private int id;
 
 	@ManyToOne
 	@JoinColumn(name="ID_CONTRAT")
@@ -52,11 +52,11 @@ public class EmploiPrecedent implements Serializable {
 		super();
 	}
 
-	public long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
